// src/types/consts.d.ts
type TagList = readonly string[];
// ...
interface Consts {
	// Start possible tags
	readonly START_POSSIBLE_TAG : TagList;
	// Unit tags for edit
	readonly UNIT_TAG : TagList;
	// Tags with a double structure
	readonly DOUBLE_TAG : TagList;
	// Tags that are not single
	readonly NOT_SINGLE_TAG : TagList;
	// Specially managed tags
	readonly SPECIAL_TAG: TagList;
	// Tags that are automatically converted by the browser during style conversion
	readonly AUTO_MODIFY_TAG : TagList;
	// Tags that allow text whitespace (regularly delete the tag if the text is whitespace)
	readonly EMPTY_ABLE_TAG : TagList;
	// Keys (id, className) of elements
	readonly CLASS_NAMES : ClassNames;
}